docs_annotation/requirements.txt
# 文档标注系统依赖

# 核心依赖
pyyaml>=6.0
pillow>=10.0.0
python-dotenv>=1.0.0

# PDF解析
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Word文档解析
python-docx>=1.0.0

# Excel解析
openpyxl>=3.1.0

# PowerPoint解析
python-pptx>=0.6.0

# HTML解析
beautifulsoup4>=4.12.0

# OCR模型（可选，按需安装）
paddleocr>=2.7.0
pytesseract>=0.3.10

# LLM接口（可选，按需安装）
openai>=1.0.0
anthropic>=0.18.0

# 高级文档解析（推荐，更高准确率）
docling>=2.0.0
